Move quarantine to utils.py
Split out quarantine into utils.py so pbench/server/__init__.py
is a bit less intimidating.
