+++ pbench/server/utils.py
@@ -1,4 +1,7 @@
 from datetime import datetime, tzinfo, timedelta
+import os
+import shutil
+import sys
 from time import time as _time
 
 
@@ -18,3 +21,29 @@ def tstos(ts=None):
         ts = _time()
     dt = datetime.utcfromtimestamp(ts).replace(tzinfo=simple_utc())
     return dt.strftime("%Y-%m-%dT%H:%M:%S-%Z")
+
+
+def quarantine(dest, logger, *files):
+    """Quarantine problematic tarballs.
+    Errors here are fatal but we log an error message to help diagnose
+    problems.
+    """
+    try:
+        os.mkdir(dest)
+    except FileExistsError:
+        # directory already exists, ignore
+        pass
+    except Exception:
+        logger.exception('quarantine {} {!r}: "mkdir -p {}/" failed', dest, files, dest)
+        sys.exit(101)
+
+    for afile in files:
+        if not os.path.exists(afile) and not os.path.islink(afile):
+            continue
+        try:
+            shutil.move(afile, os.path.join(dest, os.path.basename(afile)))
+        except Exception:
+            logger.exception(
+                'quarantine {} {!r}: "mv {} {}/" failed', dest, files, afile, dest
+            )
+            sys.exit(102)
